restructuring the project
