+++ Animal-sound/Animals/Broasca.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AnimalSound.Animals;
+
+namespace AnimalSound.Animals;
+
+public class Broasca : Animal
+{
+    public Broasca() : base("Oac Oac Oac Oaca!") { }
+}
