+++ Animal-sound/Animals/Pisica.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AnimalSound.Animals;
+
+namespace AnimalSound.Animals;
+
+public class Pisica : Animal
+{
+    public Pisica() : base("Miau Miau Miau!") { }
+}
