+++ Animal-sound/Animals/Sarpe.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using AnimalSound.Animals;
+
+namespace AnimalSound.Animals;
+
+public class Sarpe : Animal
+{
+    public Sarpe() : base("SSSSSSSSsssssssssssss!") { }
+}
