+++ Animal-sound/Display/SoundDisplay.cs
@@ -1,0 +1,11 @@
+﻿using AnimalSound.Interfaces;
+
+namespace AnimalSound.Display;
+
+public class SoundDisplay
+{
+    public void DisplaySound(IAnimal animal)
+    {
+        Console.WriteLine($"{animal.Nume} face urmatorul sunet: {animal.Sunet}");
+    }
+}
