+++ Animal-sound/Interfaces/IAnimal.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AnimalSound.Interfaces;
+
+public interface IAnimal
+{
+    string Nume { get; }
+    string Sunet { get; }
+}
